# book-network/src/main/java/tech/mouad/book/feedback/FeedBackRequest.java
package tech.mouad.book.feedback;

import jakarta.validation.constraints.*;

public record FeedBackRequest(
        @Positive(message = "200")
        @Min(value = 0, message = "201")
        @Max(value = 5, message = "202")
        double note,
        @NotNull(message = "203")
        @NotEmpty(message = "203")
        @NotBlank(message = "203")
        String comment,
        @NotNull(message = "204")
        Integer bookId
) {
}