# book-network/src/main/java/tech/mouad/book/feedback/FeedBackResponse.java
package tech.mouad.book.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedBackResponse {
    private double note;
    private String comment;
    private boolean ownFeedBack;
}